+++ packages/reakit/src/Avatar/__tests__/Avatar-test.tsx
@@ -26,6 +26,10 @@ test("styled", () => {
   color: inherit;
 }
 
+.c0:focus:not(:focus-visible) {
+  outline: none;
+}
+
 <img
   class="c0"
 />