+++ packages/reakit/src/Table/__tests__/Table-test.tsx
@@ -1,0 +1,37 @@
+import * as React from "react";
+import { render } from "react-testing-library";
+import Table from "../Table";
+
+test("html attrs", () => {
+  const { getByTestId } = render(
+    <Table id="test" aria-label="test" data-testid="test" />
+  );
+  expect(getByTestId("test")).toHaveAttribute("data-testid", "test");
+  expect(getByTestId("test")).toHaveAttribute("aria-label", "test");
+});
+
+test("styled", () => {
+  const { container } = render(<Table />);
+  expect(container.firstChild).toMatchInlineSnapshot(`
+.c0 {
+  margin: unset;
+  padding: unset;
+  border: unset;
+  background: unset;
+  font: unset;
+  font-family: inherit;
+  font-size: 100%;
+  box-sizing: border-box;
+  background-color: unset;
+  color: inherit;
+}
+
+.c0:focus:not(:focus-visible) {
+  outline: none;
+}
+
+<table
+  class="c0"
+/>
+`);
+});