+++ packages/reakit/src/_utils/applyAllRefs.ts
@@ -1,9 +1,11 @@
 import * as React from "react";
 
 function applyAllRefs<T = any>(...refs: Array<React.Ref<T> | undefined>) {
+  const validRefs = refs.filter(Boolean);
+  if (!validRefs.length) return undefined;
   return (element: T) => {
-    for (const value of refs) {
-      if (typeof value === "object" && value !== null) {
+    for (const value of validRefs) {
+      if (typeof value === "object") {
         // @ts-ignore
         value.current = element;
       } else if (typeof value === "function") {