+++ rollup.config.js
@@ -37,7 +37,7 @@ const getPlugins = umd =>
         commonjs({
           include: /node_modules/,
           namedExports: {
-            "../../node_modules/styled-components/node_modules/react-is/index.js": [
+            "../../node_modules/react-is/index.js": [
               "isValidElementType",
               "isElement",
               "ForwardRef"